Configurations benchmark (#109)
* can set configuration.py

* global debug

* global debug

* add end time sqlite

* add run id to get all runs sqlite

* remove print

* add cpufreq to system samples

* add cpufreq to system samples

* system service is getting cpu freq

* fixing loading annimation

* sleep interval to 3

* update slurm api

* version 3

* publish script

* version 0.3.1

* permissions to write to local storage

* full path

* abselute path

* fix permissionerror handling

* codestyle

* version 0.3.5

* version 0.3.6

* no disabled

+++ chronus/domain/cpu_info.py
@@ -14,7 +14,7 @@ class SystemInfo:
     cpu_name: str = ""
     cores: int = 0
     threads_per_core: int = 1
-    frequencies: list[float] = None
+    frequencies: list[int] = None
 
     def __hash__(self):
         return hash((self.cpu_name, self.cores, self.threads_per_core, tuple(self.frequencies)))
